rename input to inputfield and container to input
